cleanup + reducing bundle size further
