+++ src/utils/decodeUserJwt.ts
@@ -1,4 +1,4 @@
-import { debug } from "debug";
+import { debug } from "./logger";
 import { User } from "../types";
 import { b64decode } from "./cookieUtils";
 const _log = debug("squatch-js:decodeUserJwt");
